requirements.txt
django
phonenumbers
django-phonenumber-field
django_rest_framework
